% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEveScores.R
\name{getEveScores}
\alias{getEveScores}
\title{Gets the EVE score for the amino acids in variant of interest.}
\usage{
getEveScores(eveData, variantData, protein = TRUE)
}
\arguments{
\item{eveData}{A tibble containing the EVE data that has already been
processed by the processEveData function.}

\item{variantData}{A tibble containing the variant data that has already
been processed by the processVariantData function.  Assumption is that there
will only be one variant at each residue/genomic position and the wildtype
and variant amino acid are distinct from one another.  For protein data it
will be missense variants only and genomic data single nucleotide variants
only.  Won't try to score mutations that result in wildtype and variant amino
acid being the same.  Assume variants are independent of each other.}

\item{protein}{Specifies whether the variant data is in protein or genomic
form.  By default it is set to TRUE meaning the default is protein form.}
}
\value{
Returns a tibble containing the EVE scores for each residue
position that has a score calculated by EVE, residue position, wildtype amino
acid, and mutated amino acid.  If there are NaNs it means the variant
provided doesn't have an EVE score.  The names represent the residue
position.
}
\description{
A function that takes in EVE and variant data then will get the EVE score for
each amino acid in the variant of interest.  If the data is supplied in
protein form it will take the average of the EVE scores for that amino acid
mutation since EVE data is available down to the single nucleotide variation
level and multiple codons code for the same amino acid.
}
