% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processVariantData.R
\name{processVariantData}
\alias{processVariantData}
\title{Processes variant data in csv format}
\usage{
processVariantData(filePath, protein = TRUE)
}
\arguments{
\item{filePath}{The filepath to the csv file containing variant information
about the variant and gene of interest.}

\item{protein}{This argument specifies whether the input data is in genomic
or protein coordinates.}
}
\value{
Returns a tibble with the data from the csv file provided.
}
\description{
A function that checks if the data provided is in the right format (csv), if
it has the columns required for downstream analysis depending on if
the variant data is in genomic or protein coordinates, and processes it for
future use.
}
