% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runVariantMapper.R
\name{runVariantMapper}
\alias{runVariantMapper}
\title{Launch Shiny App for variantMapper}
\usage{
runVariantMapper()
}
\value{
No return value but will open up a Shiny page.
}
\description{
A function that launches the Shiny app for variantMapper, the code for the
Shiny app haas been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{
variantMapper::runVariantMapper()
}

}
