% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreVariant.R
\name{scoreVariant}
\alias{scoreVariant}
\title{Scores a variant of a protein based on weighted average.}
\usage{
scoreVariant(
  eveScores,
  posWeights = rep(1/length(eveScores), length(eveScores))
)
}
\arguments{
\item{eveScores}{EVE scores for each residue position that has a possible EVE
score.}

\item{posWeights}{The weights to give each residue position when calculating
the weighted average of EVE scores for the protein variant.  The default is
using equal weights for each residue position i.e. 1/(number of residues with
EVE scores).  This is used when an argument is not passed into the posWeights
parameter.}
}
\value{
Returns a double with the weighted EVE score for the protein variant.
}
\description{
Scores a variant of a protein based on weighted average.
}
