% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreVariants.R
\name{scoreVariants}
\alias{scoreVariants}
\title{Scores the variant of interest with EVE scores}
\usage{
scoreVariants(eveData, variantData, protein = TRUE)
}
\arguments{
\item{eveData}{A tibble containing the EVE data that has already been
processed by the processEveData function.}

\item{variantData}{A tibble containing the variant data that has already
been processed by the processVariantData function.  Assumption is that there
will only be one variant at each residue/genomic position and the wildtype
and variant amino acid are distinct from one another.  For protein data it
will be missense variants only and genomic data single nucleotide variants only.
Won't try to score mutations that result in wildtype and variant amino acid as the same
Assume variants are independent of each other, and only 1 variant at each genomic/residue position.}

\item{protein}{Specifies whether the variant data is in protein or genomic
form.  By default it is set to TRUE meaning the default is protein form.}
}
\value{
Returns a vector containing the EVE scores for each residue position
that has a score calculated by EVE.  If there are NaNs it means the
variant provided doesn't have an EVE score.
}
\description{
A function that takes in EVE and variant data then will score the variant
with an EVE score.  If the data is supplied in protein form it will take the
average of the EVE scores for that amino acid mutation since EVE data is
available down to the single nucleotide variation level.
}
