% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeVariant.R
\name{visualizeVariant2}
\alias{visualizeVariant2}
\title{Function to visualize EVE scores of two variants at once}
\usage{
visualizeVariant2(eveInfo1, eveInfo2)
}
\description{
Function to visualize EVE scores of two variants at once
}
