% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processVariantData.R
\name{processVariantData}
\alias{processVariantData}
\title{Processes variant data in csv format}
\usage{
processVariantData(filePath, protein = TRUE)
}
\arguments{
\item{filePath}{A character vector of the filepath to the csv file containing
variant information for the gene of interest.}

\item{protein}{A logical value specifying whether the input data is in
genomic or protein coordinates.  If TRUE the data contains protein
information and protein coordinates; this is the default.  If FALSE it
contains genomic information and genomic coordinates.}
}
\value{
A tibble with the data from the csv file provided.
}
\description{
A function that checks if the data provided is in the right format (csv), if
it has the columns required for downstream analysis depending on if
the variant data is in genomic or protein coordinates, and processes it for
future use.
}
\examples{
# Example
# Get the variant data in the right format depending on if it is protein or
# genomic data.

# If the data is in protein form.
varDataProtPath <- system.file("extdata", "variant_data_protein.csv", package = "variantMapper")
varDataProt <- processVariantData(varDataProtPath, protein = TRUE)
varDataProt

# If the data is in genomic form.
varDataGenPath <- system.file("extdata", "variant_data_genomic.csv", package = "variantMapper")
varDataGen <- processVariantData(varDataGenPath, protein = FALSE)
varDataGen

}
