% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeVariant.R
\name{visualizeVariant}
\alias{visualizeVariant}
\title{Visualizes EVE score for a protein variant.}
\usage{
visualizeVariant(eveInfo)
}
\arguments{
\item{eveInfo}{Tibble with EVE scores for each residue
position that has a score calculated by EVE, residue position, wildtype amino
acid, and mutated amino acid.  If there are NaNs it means the variant
provided doesn't have an EVE score.}
}
\value{
A lollipop graph showing the EVE score at each residue position.
}
\description{
A function that visualizes EVE scores of a variant at residue positions where
EVE scores are possible.
}
